requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
boto3==1.40.13
botocore==1.40.13
Bottleneck==1.4.2
Brotli==1.0.9
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
cloudpickle==3.1.1
contourpy==1.2.1
cryptography==45.0.6
cycler==0.11.0
databricks-sdk==0.63.0
Deprecated==1.2.18
dnspython==2.7.0
docker==7.1.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
Flask==3.1.2
fonttools==4.55.3
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.4.0
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.4
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
minio==7.2.16
mlflow==3.1.4
mlflow-skinny==3.1.4
multidict==6.6.3
numexpr==2.10.1
numpy==2.0.2
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==2.3.1
paramiko==4.0.0
pillow==11.3.0
pip==25.3
prometheus_client==0.22.1
prometheus_flask_exporter==0.23.2
propcache==0.3.1
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.13.2
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyu2f==0.1.5
PyYAML==6.0.2
querystring_parser==1.2.4
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rsa==4.9.1
s3transfer==0.13.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
setuptools==72.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.2
tenacity==9.1.2
threadpoolctl==3.5.0
tomli==2.2.1
tornado==6.5.1
typer==0.16.1
typer-slim==0.16.1
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
unicodedata2==15.1.0
urllib3==1.26.19
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.21.0
zstandard==0.23.0
